Wipe the BL's magic number before entering the BL
The magic number that we use to indicate to ourselves that we should
enter the bootloader on reset doesn't get overwritten by the
bootloader.  This is problematic, as once the bootloader has done its
job, we immediately enter the bootloader again.

So we should wipe the magic number just before we enter the bootloader
to ensure that we don't get stuck.

This patch also changes the magic number.  This ensures that if we are
updating from a previous version of this firmware then we're not going
to go straight into the bootloader again.